Removes useless abstract types
