+++ NachoSpigot-Server/src/main/java/com/github/sadcenter/auth/utils/HashMapBuilder.java
@@ -1,4 +1,4 @@
-package com.github.sadcenter.impl.utils;
+package com.github.sadcenter.auth.utils;
 
 import java.util.HashMap;
 import java.util.Map;
